Move Ask blocks to remove circular ref
