+++ cfgov/cfgov/settings/base.py
@@ -613,9 +613,11 @@ FLAGS = {
     # Supports testing of a new 2021 version of the website home page.
     # Enable by appending ?home_page_2021=True to home page URLs.
     "HOME_PAGE_2021": [
+        ("environment is not", "production", True),
         ("parameter", "home_page_2021"),
     ],
     "HOME_PAGE_2021_MINIMAL": [
+        ("environment is not", "production", True),
         ("parameter", "home_page_2021_minimal"),
     ],
     # Whether robots.txt should block all robots, except for Search.gov.
